Camera: Loosen restrictions on readoutTimestamp.
Camera Framework currently forces all AIDL HALs to support
readoutTimestamp, which means HALs that don't have this capability
have to fake the readout timestamp.

However, considering readoutTimestamp key is intended to be
actual information coming from the sensor, this CL loosens the
restriction and allows all HALs to advertise NOT_SUPPORTED if
the HAL is incapable of getting the sensor's readout timestamp
(External Camera HALs, for example).

This CL also changes the visibility of the key to be available to
the HAL and the applications over Java and NDK.

Bug: 309543399
Test: Generated docs.html looks OK
Change-Id: Ie6bb82850254b0c4e5845d30ff2d6303f566a203