Camera: Update metadata generation scripts to make copyright year sticky
The scripts that take generate metadata related file had a
hardcoded copyright year that would get applied to all files
regardless of of if the files previously existed or not. This
value was set to 2022.

Because metadata keys may be added in a year that is not 2022,
this CL updates the scripts to use the current year when creating
new files and read the existing files to determine the correct
copyright year for existing files.

Bug: n/a
Test: Manually verified that re-running metadata-generate with
      a different copyright year does not create unnecessary diffs.
Change-Id: I97c7138b0188209342b0f89d8b9b4b77f6f57c29