camera: Add session_characteristics_key_since
Certain camera characteristics keys can now be modified according to
session configurations. The camera characteristics that have values
modified to account for some session configuration are called
session characteristics, and the keys that can differ between camera
characteristics and session characteristics are limited and statically
known.

As the list of keys that differ between camera characteristics and
session characteristics are statically known for a given API level,
this CL updates metadata_definitions to accept a new
`session_characteristics_key_since` attribute which is the source of
truth for the API level after which a particular key could differ.

As this information is required by camera2 and cameraserver, the
generation scripts were updated to encode this information in
CameraCharacteristics.java and SessionCharacteristicsTags.h

Bug: 303645857
Test: metadata-generate runs fine and doesn't cause unnecessary
      edits.
Test: atest android.hardware.camera2.cts.CameraDeviceSetupTest passes
Change-Id: If9bc534ef03d98d0f16433226290ca6d15a41d63