Add DeviceffectHandle mutex
Device Effects can be controlled by the framework itself each
time routing is changed. (checking if the port associated to the effect
is involved or not in an audio patch to enable/disable this effect).
This specific device effect handle follows a different sequence than
regular effect, which are expected to be controled by an effect
application. Hence the handle mutex is expected to be on the top of the
mutex chain.

Flag: EXEMPT bugfix
Bug: 329395147
Bug: 348986455
Test: atest CtsMediaAudioTestCases

Change-Id: I750707b72707d9704e14f3829e170d3173605a87
Signed-off-by: François Gaffie <francois.gaffie@ampere.cars>